Give less energy from photosyntesis
It was stupid that grass had to prevent its own growth to limit its numbers.
I've changed photosynthesis to give less energy the more plants there are in the
area... this is supposed to simulate how plants compete for light. Probably in
the future it will make sense to give some benefits to bigger plants.

As a plus, this change allowed me to simplify algorithm of grass. Minus is a
significant performance drop.