Reuse result table in creature_look
Profiler indicated that creating new table each time creature look is called is
costly. This is because the table is filled from empty to N, causing a lot of
rehashing (allocated size is increased *2 every time). By reusing this table, it
will grow and then remain "big enough".

Profiler showed a massive improvement... but when the code runs without the
profiler, I see no gain.

I decided to deliver it because I do use the profiler, and it runs faster that
way.