Avoid the need to sort results of creature_look
Results of creature_look are built based on the content of a grid, which stores
creates in a hash set that does not prevent the order. This causes indeterminism
in tournaments. To avoid that, I've been sorting the results.

Profiler showed that it is faster to use order preserving arrays to implement
the grid than to sort those results, and I did that.