Require reaching turn limit to pass qualifications
Fixing a bug that declared species as successfully passing qualification even
when it didn't survive till the turn limit.

Now it works as was my original intention. If it goes extinct too soon, it
fails. This ways creatures are forced to eat (by a high turn limit) but cannot
eat everything (starvation risk).