Reduce energy based on amounts of computation
In order to kill creatures that may start executing an endless loop as well as
to encourage effective algorithms, I'm introducing a penalty for executing
instructions.

So far, running the brain had no cost, apart from the fixed turn cost and the
cost associated with made decisions. From now on, the longer it takes to execute
brain, the more energy it will cost.

I don't want to kill creativity or discourage complicated algorithms, so the
cost is very low. I'm not sure about the value, but I set it so that the longest
execution of a brain that I've seen in 1st edition will cost just 0.1 energy
points.