Disallow breeding with negative offspring's energy
This was an interesting bug letting you gain infinite amounts of energy and
breed forever. For a long time I was tempted to leave it, but finally decided to
fix it, because exploiting it could cause performance issues (too many living
beings in the game).

So now you'll die if you try to breed with a negative value of energy for the
offspring.