Simplify game finishing logic
Game used to be 'finished' when there was only one living species left. That
logic was not very accurate with my intentions, and the game-ending conditions
are better handled in main, as they depend on the mode (tournament vs
qualification).

This commit simplifies the code in game. While there is something living, the
game keeps running. Even if all living creatures are of the same species. This
way a player can be qualified if it manages to survive till the turn limit, even
if everything else died a few turns before.