Improve inconsistent naming of species' names
As reported in issue #2, there was an inconsistency related to how a species is
named. Players were providing their name via 'species.name', but referred to
names of other creatures as 'creature.species'.

I've changed the code so now the names of the species of other creatures are
accessible under 'creature.name'