Get rid of trees
I've decided that the presence of a few trees could lead to unbalanced game,
because player who is lucky and spawns near a tree has an easy access to a huge
deposit of energy. To make this balanced, I would need way more trees, and that
would require much higher starting energies... Something that is doable but I
don't want that yet.